requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
python-decouple
gunicorn
africastalking
django-oidc-provider
pytest
pytest-django
coverage
python-dotenv
